fixed registerPostgresCommands to match delete account PR

+++ src/postgres/commands/deletePostgresDatabase.ts
@@ -3,13 +3,17 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { IActionContext } from "vscode-azureextensionui";
+import { DialogResponses, IActionContext } from "vscode-azureextensionui";
 import { ext } from "../../extensionVariables";
 import { PostgresDatabaseTreeItem } from "../tree/PostgresDatabaseTreeItem";
 
-export async function deleteDatabase(context: IActionContext, node?: PostgresDatabaseTreeItem): Promise<void> {
+export async function deletePostgresDatabase(context: IActionContext, node?: PostgresDatabaseTreeItem): Promise<void> {
     if (!node) {
         node = <PostgresDatabaseTreeItem>await ext.tree.showTreeItemPicker(PostgresDatabaseTreeItem.contextValue, context);
     }
-    await node.deleteTreeItem(context);
+    const message: string = `Are you sure you want to delete database "${node.databaseName}"?`;
+    const result = await ext.ui.showWarningMessage(message, { modal: true }, DialogResponses.deleteResponse);
+    if (result === DialogResponses.deleteResponse) {
+        await node.deleteTreeItem(context);
+    }
 }
